LangChain/requirements.txt
langchain
langchain-openai
langchain-community
langchain-experimental
pypdf
faiss-cpu
python-dotenv
pydantic
torch
transformers
pdf2image
PyMuPDF
tqdm
timm